try to improve untracked files situation
hardware-configuration.nix - Minimal example from repos that just
imports not-detected.nix.  Turns out the error messaging from
nixos-rebuild is actually pretty good so this isn't an issue.

secrets.nix - Remove file altogether and source it from a local box.
This has the disadvantage that initial setup requires LAN access, but
that's better than the alternative of manually having to source the file
for setup.