umu: inject extra libs into compatibility tool dir with overlayfs
Unfortunately, while hacky, this is the only way to get system libraries
back into Proton.  Proton is opinionated about the libraries passed
through to games and explicitly sets LD_LIBRARY_DIR to point only to
these libraries, so even if UMU were to implement support for multiPkgs
in the FHS, this would have no effect.